# api-docs/swagger.json
{
    "swagger":"2.0",
    "info":{
      "version":"1.0.0",
      "title":"Money tracker"
    },
    "host":"ec2-35-181-53-65.eu-west-3.compute.amazonaws.com:3000",
    "basePath":"/",
...
    "schemes":[
      "http"
    ],
    "paths":{
...
      "/charges":{
        "post":{
          "tags":[
            "charge"
          ],
          "summary":"Create income or charge entry",
          "description":"",
          "operationId":"createCharge",
          "consumes":[
            "application/json"
          ],
          "produces":[
            "application/json",
            "application/xml"
          ],
          "parameters":[
            {
              "in":"body",
              "name":"body",
              "required":true,
              "schema":{
                "$ref":"#/definitions/createCharge"
              }
            }
          ],
          "responses":{
            "200":{
              "description":"successful operation"
            },
            "400":{
              "description":"Invalid Entry"
            }
          }
        },
        "get":{
          "tags":[
            "charge"
          ],
          "summary":"Find all entry",
          "operationId":"getCharges",
          "produces":[
            "application/json",
            "application/xml"
          ],
          "parameters":[
            {
              "name":"type",
              "in":"query",
              "description":"type of charge ('charge' | 'income')",
              "required":true,
              "type":"string"
            },
            {
              "name":"from",
              "in":"query",
              "description":"start date (in ms)",
              "required":false,
              "type":"number"
            }
          ],
          "responses":{
            "200":{
              "description":"successful operation"
            }
          }
        },
        "put":{
          "tags":[
            "charge"
          ],
          "summary":"Edit specific entry",
          "operationId":"editCharge",
          "produces":[
            "application/json",
            "application/xml"
          ],
          "parameters":[
            {
              "in":"body",
              "name":"body",
              "required":true,
              "schema":{
                "$ref":"#/definitions/editCharge"
              }
            }
          ],
          "responses":{
            "200":{
              "description":"successful operation"
            },
            "400":{
              "description":"Invalid Entry"
            }
          }
        },
        "delete":{
          "tags":[
            "charge"
          ],
          "summary":"Delete charge by ID",
          "operationId":"deleteCharge",
          "produces":[
            "application/json",
            "application/xml"
          ],
          "parameters":[
            {
              "name":"id",
              "in":"query",
              "description":"ID of the charge that needs to be deleted",
              "required":true,
              "type":"integer",
              "minimum":1,
              "format":"int64"
            }
          ],
          "responses":{
            "200":{
              "description":"successful operation"
            },
            "400":{
              "description":"Invalid ID supplied"
            },
            "404":{
              "description":"Charge not found"
            }
          }
        }
      },
      "/categories":{
        "get":{
          "tags":[
            "category"
          ],
          "summary":"Find all categories",
          "operationId":"getCategories",
          "produces":[
            "application/json",
            "application/xml"
          ],
          "parameters":[
  
          ],
          "responses":{
            "200":{
              "description":"successful operation"
            }
          }
        },
        "delete":{
          "tags":[
            "category"
          ],
          "summary":"Delete category by ID",
          "operationId":"deleteCategory",
          "produces":[
            "application/json",
            "application/xml"
          ],
          "parameters":[
            {
              "name":"id",
              "in":"query",
              "description":"ID of the order that needs to be deleted",
              "required":true,
              "type":"integer",
              "minimum":1,
              "format":"int64"
            }
          ],
          "responses":{
            "200":{
              "description":"successful operation"
            },
            "400":{
              "description":"Invalid ID supplied"
            },
            "404":{
              "description":"Category not found"
            }
          }
        },
        "put":{
          "tags":[
            "category"
          ],
          "summary":"Edit specific category",
          "operationId":"editCategory",
          "produces":[
            "application/json",
            "application/xml"
          ],
          "parameters":[
            {
              "in":"body",
              "name":"body",
              "required":true,
              "schema":{
                "$ref":"#/definitions/editCategory"
              }
            }
          ],
          "responses":{
            "200":{
              "description":"successful operation"
            },
            "400":{
              "description":"Invalid Entry"
            }
          }
        },
        "post":{
          "tags":[
            "category"
          ],
          "summary":"Create new category",
          "operationId":"createCategory",
          "produces":[
            "application/json",
            "application/xml"
          ],
          "parameters":[
            {
              "in":"body",
              "name":"body",
              "required":true,
              "schema":{
                "$ref":"#/definitions/createCategory"
              }
            }
          ],
          "responses":{
            "200":{
              "description":"successful operation"
            },
            "400":{
              "description":"Invalid Entry"
            }
          }
        }
      },
...
    },
    "definitions":{
...
      "createCategory": {
        "type":"object",
        "required":[
          "name",
          "icon"
        ],
        "properties":{
          "name":{
            "type":"string"
          },
          "description":{
            "type":"string"
          },
          "iconId":{
            "type":"integer",
            "format":"int64"
          }
        },
        "xml":{
          "name":"createCategory"
        }
      },
      "editCategory": {
        "type":"object",
        "required":[
            "id",
          "name",
          "iconId"
        ],
        "properties":{
          "id":{
            "type":"integer",
            "format":"int64"
          },
          "name":{
            "type":"string"
          },
          "description":{
            "type":"string"
          },
          "iconId":{
            "type":"integer",
            "format":"int64"
          }
        },
        "xml":{
          "name":"editCategory"
        }
      },
...
      "createCharge": {
        "type":"object",
        "required":[
          "name",
          "icon",
          "categoryId"
        ],
        "properties":{
          "categoryId":{
            "type":"integer",
            "format":"int64"
          },
          "description":{
            "type":"string"
          },
          "date": {
            "type":"integer",
            "format":"int64"
          },
          "money":{
            "type":"integer",
            "format":"int64"
          },
          "type":{
            "type":"string",
            "description":"charge type in the db",
            "enum":[
              "charge",
              "income"
            ]
          }
        },
        "xml":{
          "name":"createCharge"
        }
      },
      "editCharge": {
        "type":"object",
        "required":[
            "id",
          "name",
          "iconId",
          "type",
          "money",
          "categoryId"
        ],
        "properties":{
          "id":{
            "type":"integer",
            "format":"int64"
          },
          "categoryId":{
            "type":"integer",
            "format":"int64"
          },
          "description":{
            "type":"string"
          },
          "date": {
            "type":"integer",
            "format":"int64"
          },
          "money":{
            "type":"integer",
            "format":"int64"
          },
          "type":{
            "type":"string",
            "description":"charge type in the db",
            "enum":[
              "charge",
              "income"
            ]
          }
        },
        "xml":{
          "name":"editCharge"
        }
      },
...
    }
  }